+++ 1-EntradaSalida/jsEntradaSalida-10.js
@@ -5,7 +5,7 @@ mostrar el importe con un Descuento del 25 %
 en el cuadro de texto "RESULTADO"*/
 function mostrarAumento()
 {
-	let importe=parseInt(document.getElementById("txtIdImporte").value);
+	let importe=parseFloat(document.getElementById("txtIdImporte").value);
 	let porcentaje= 0.25*importe;
 	let descuento= importe-porcentaje;
 	document.getElementById("txtIdResultado").value= descuento;
